Fix oob get_unchecked, overlapping _nonoverlapping
In quite a few places this code wants a pointer into a slice, and it
gets that pointer by calling get_unchecked. This is potentially valid,
but in general extremely dangerous because get_unchecked is a slice
method even when called on a Vec, so regardless of whether the pointer
would point to valid memory, if the index is out of bounds of the slice,
that's UB. Offsetting pointers provides the less surprising semantics.